# specs/004-cacao-processing-module/contracts/openapi.yml
openapi: 3.0.0
info:
  title: Cacao Processing API
  version: 1.0.0
  description: API for managing cacao processing batches and laboratory samples.
paths:
  /api/batches:
    get:
      summary: Get all cacao batches
      responses:
        '200':
          description: A list of cacao batches.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CacaoBatch'
    post:
      summary: Create a new cacao batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCacaoBatch'
      responses:
        '201':
          description: The created cacao batch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacaoBatch'
  /api/batches/{batchId}:
    get:
      summary: Get a single cacao batch
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A single cacao batch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacaoBatch'
    put:
      summary: Update a cacao batch with final dried weight
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCacaoBatch'
      responses:
        '200':
          description: The updated cacao batch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacaoBatch'
  /api/receptions/{receptionId}/samples:
    get:
      summary: Get all laboratory samples for a reception
      parameters:
        - name: receptionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A list of laboratory samples.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LaboratorySample'
    post:
      summary: Create a new laboratory sample for a reception
      parameters:
        - name: receptionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLaboratorySample'
      responses:
        '201':
          description: The created laboratory sample.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LaboratorySample'
  /api/samples/{sampleId}:
    get:
      summary: Get a single laboratory sample
      parameters:
        - name: sampleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A single laboratory sample.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LaboratorySample'
    put:
      summary: Update a laboratory sample with quality results
      parameters:
        - name: sampleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLaboratorySample'
      responses:
        '200':
          description: The updated laboratory sample.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LaboratorySample'
components:
  schemas:
    CacaoBatch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        batch_type:
          type: string
        start_date:
          type: string
          format: date-time
        duration:
          type: integer
        total_wet_weight:
          type: number
        total_dried_weight:
          type: number
        status:
          type: string
        expected_completion_date:
          type: string
          format: date-time
        total_sacos_70kg:
          type: integer
        remainder_kg:
          type: number
    CreateCacaoBatch:
      type: object
      properties:
        reception_ids:
          type: array
          items:
            type: string
            format: uuid
        batch_type:
          type: string
        start_date:
          type: string
          format: date-time
        duration:
          type: integer
    UpdateCacaoBatch:
      type: object
      properties:
        total_sacos_70kg:
          type: integer
        remainder_kg:
          type: number
    LaboratorySample:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reception_id:
          type: string
          format: uuid
        sample_weight:
          type: number
        estimated_drying_days:
          type: integer
        status:
          type: string
        dried_sample_kg:
          type: number
        violetas_percentage:
          type: number
        moho_percentage:
          type: number
        basura_percentage:
          type: number
    CreateLaboratorySample:
      type: object
      properties:
        sample_weight:
          type: number
        estimated_drying_days:
          type: integer
    UpdateLaboratorySample:
      type: object
      properties:
        dried_sample_kg:
          type: number
        violetas_percentage:
          type: number
        moho_percentage:
          type: number
        basura_percentage:
          type: number